+++ src/wetwire_github/actions/__init__.py
@@ -2,10 +2,16 @@
 
 from .cache import cache
 from .checkout import checkout
+from .codecov import codecov
+from .configure_aws_credentials import configure_aws_credentials
+from .create_pull_request import create_pull_request
 from .docker_build_push import docker_build_push
 from .docker_login import docker_login
+from .docker_metadata import docker_metadata
 from .download_artifact import download_artifact
+from .gh_release import gh_release
 from .github_script import github_script
+from .setup_buildx import setup_buildx
 from .setup_dotnet import setup_dotnet
 from .setup_go import setup_go
 from .setup_java import setup_java
@@ -17,10 +23,16 @@ from .upload_artifact import upload_artifact
 __all__ = [
     "cache",
     "checkout",
+    "codecov",
+    "configure_aws_credentials",
+    "create_pull_request",
     "docker_build_push",
     "docker_login",
+    "docker_metadata",
     "download_artifact",
+    "gh_release",
     "github_script",
+    "setup_buildx",
     "setup_dotnet",
     "setup_go",
     "setup_java",